healthcare-agent-backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
azure-ai-formrecognizer
azure-search-documents
azure-identity
aiortc
requests
azure-cognitiveservices-speech
azure-storage-blob
python-multipart
gunicorn
semantic-kernel
azure-ai-textanalytics
azure-cosmos
azure-communication-email
python-jose[cryptography]